refactor: move SwapForm into a separate file
